+++ src/components/page/HR.vue
@@ -282,6 +282,8 @@
                         console.log(this.hr_list)
                     }else{
                         self.$message.error('暂无数据！')
+                        self.totalNumber=0
+                        self.hr_list=[];
                     }
                 })
                 self.cur_page=1;
@@ -311,20 +313,29 @@
                     this.options4 = [];
                 }
             },
-            searchCompany(){
-
-            },
             deleteHR(id){
                 this.deleteId=id;
                 let self = this;
                 self.url = DELETEHR;
-                self.$axios.delete(self.url+this.deleteId).then((response) => {
-                    console.log(response);
-                    if(response.data == ''){
-                        this.getData();
-                        self.$message.error('删除成功！')
-                    }
-                })
+                this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
+                    confirmButtonText: '确定',
+                    cancelButtonText: '取消',
+                    type: 'warning',
+                    center: true
+                }).then(() => {
+                    self.$axios.delete(self.url+this.deleteId).then((response) => {
+                        console.log(response);
+                        if(response.data == ''){
+                            this.getData();
+                            self.$message.error('删除成功！')
+                        }
+                    })
+                }).catch(() => {
+                    this.$message({
+                        type: 'info',
+                        message: '已取消删除'
+                    });
+                });
             },
             modifyManageplatform(id){
                 this.modifyId=id;
